pyproject.toml
[tool.poetry]
name = "test-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Dontgoingforyou <mishael000@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
python-telegram-bot = "21.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
